+++ demo/views/modals/modals.js
@@ -1,5 +1,8 @@
 import React from 'react';
 import DialogDemo from './dialog-demo';
+import SidebarDemo from './sidebar-demo';
+import AlertDemo from './alert-demo';
+import ConfirmDemo from './confirm-demo';
 
 class Modals extends React.Component {
   /**
@@ -10,6 +13,9 @@ class Modals extends React.Component {
       <div>
         <h1>Modals</h1>
         <DialogDemo />
+        <SidebarDemo />
+        <AlertDemo />
+        <ConfirmDemo />
       </div>
     );
   }